Split the installer into layers and gate the analysis scripts

+++ fulcrum/ui/ui_scale.py
@@ -20,4 +20,4 @@ def factor() -> float:
 
 
 def px(value: float) -> int:
-    return int(round(value * _state["factor"]))
+    return round(value * _state["factor"])
